refactor(ssm): split the SSM op kernels by recurrence
`megatron/core/ssm/ops/` was a flat directory holding the Mamba2 SSD kernels
alongside the handful of kernels that are not specific to any recurrence. That
was tenable while Mamba2 was the only variant with forked kernels. It stops
being tenable as soon as a second variant lands its own fork, because there is
then no way to tell from the layout which kernels a given mixer owns.

Split into two packages:

* `ops/common/` -- the short causal convolution (`causal_conv1d_triton`,
  `causal_conv1d_varlen`), the slot-indexed state gather/scatter that prefix
  caching uses (`intermediate_extraction`), and the autotune-config filtering
  that keeps the kernels deterministic (`determinism`). Every linear-attention
  mixer needs these; `determinism` was already shared between the SSD kernels
  and the varlen convolution.
* `ops/mamba2/` -- the state-space duality kernels (`ssd_*`, `mamba_ssm`,
  `batch_invariant_decode`).

A later variant with its own forked kernels adds a sibling package next to
`mamba2` rather than growing the flat directory.

`intermediate_extraction` reads as Mamba-specific from its docstring, but the
kernels are a plain conditional gather/scatter over flat state rows with no
Mamba math in them, so it belongs with the shared code.

Pure move. The top-level `ops` re-exports are unchanged, so callers that go
through `megatron.core.ssm.ops` need no edit. Unit tests move to mirror the
new layout.
