+++ megatron/core/ssm/ops/common/causal_conv1d_varlen.py
@@ -10,7 +10,7 @@ import torch
 import triton
 import triton.language as tl
 
-from megatron.core.ssm.ops.determinism import autotune_configs
+from megatron.core.ssm.ops.common.determinism import autotune_configs
 
 
 # Two block-dim regimes:
