+++ megatron/core/ssm/ops/mamba2/batch_invariant_decode.py
@@ -7,10 +7,10 @@ import torch
 import triton
 import triton.language as tl
 
-from megatron.core.ssm.ops.ssd_bmm import _bmm_chunk_fwd
-from megatron.core.ssm.ops.ssd_chunk_scan import _chunk_scan_fwd
-from megatron.core.ssm.ops.ssd_chunk_state import _chunk_cumsum_fwd, _chunk_state_fwd
-from megatron.core.ssm.ops.ssd_state_passing import _state_passing_fwd
+from megatron.core.ssm.ops.mamba2.ssd_bmm import _bmm_chunk_fwd
+from megatron.core.ssm.ops.mamba2.ssd_chunk_scan import _chunk_scan_fwd
+from megatron.core.ssm.ops.mamba2.ssd_chunk_state import _chunk_cumsum_fwd, _chunk_state_fwd
+from megatron.core.ssm.ops.mamba2.ssd_state_passing import _state_passing_fwd
 
 
 @triton.jit
