+++ megatron/core/ssm/ops/mamba2/ssd_chunk_scan.py
@@ -7,7 +7,7 @@ import triton
 import triton.language as tl
 from packaging import version
 
-from megatron.core.ssm.ops.determinism import autotune_configs
+from megatron.core.ssm.ops.common.determinism import autotune_configs
 
 TRITON_22 = version.parse(triton.__version__) >= version.parse("2.2.0")
