+++ megatron/core/ssm/ops/mamba2/ssd_chunk_state.py
@@ -8,7 +8,7 @@ import triton
 import triton.language as tl
 from packaging import version
 
-from megatron.core.ssm.ops.determinism import autotune_configs
+from megatron.core.ssm.ops.common.determinism import autotune_configs
 
 try:
     TRITON3 = version.parse(triton.__version__) >= version.parse("3.0.0")
