+++ megatron/core/ssm/ops/mamba2/ssd_state_passing.py
@@ -7,7 +7,7 @@ import torch
 import triton
 import triton.language as tl
 
-from megatron.core.ssm.ops.determinism import autotune_configs
+from megatron.core.ssm.ops.common.determinism import autotune_configs
 
 
 @triton.autotune(
