+++ tests/unit_tests/ssm/ops/mamba2/test_ssd_bmm.py
@@ -5,7 +5,7 @@ import unittest
 import torch
 
 try:
-    from megatron.core.ssm.ops.ssd_bmm import _bmm_chunk_fwd
+    from megatron.core.ssm.ops.mamba2.ssd_bmm import _bmm_chunk_fwd
 
     HAVE_SSD_OPS = True
 except (ImportError, Exception):
